+++ lib/index.dart
@@ -1,12 +1,11 @@
 export 'package:device_preview/device_preview.dart' show DevicePreview;
-export 'package:flutter/foundation.dart';
+
 export 'package:flutter/material.dart';
 export 'package:provider/provider.dart';
 
 /* screens */
-export 'screens/topbar/topbar.dart';
-export 'screens/explorer/explorer.dart';
+export 'screens/main/mainPage.dart';
 
 /* providers */
-export 'providers/explorer.dart';
 export 'providers/common.dart';
+export 'providers/explorer.dart';
