+++ lib/providers/common.dart
@@ -2,9 +2,14 @@ import 'package:flutter/foundation.dart';
 
 class CommonProvider with ChangeNotifier, DiagnosticableTreeMixin {
   late double bodyHeight;
+  late double screenWidth;
 
   void setBodyHeight(double bodyHeight) {
     this.bodyHeight = bodyHeight;
+  }
+
+  void setScreenWidth(double screenWidth) {
+    this.screenWidth = screenWidth;
   }
 
   @override
